cgi-bin/requirements.txt
Flask>=2.0.0,<3.0.0
PyJWT>=2.0.0
Flask-APScheduler>=1.13.0
Flask-Sqlalchemy
Flask-Login
PyMySQL
oauthlib